+++ src/Commands/MakeValueObject.php
@@ -3,15 +3,11 @@
 namespace Lunarstorm\LaravelDDD\Commands;
 
 use Illuminate\Console\Command;
+use Symfony\Component\Console\Input\InputArgument;
 
 class MakeValueObject extends DomainGeneratorCommand
 {
-    /**
-     * The name and signature of the console command.
-     *
-     * @var string
-     */
-    protected $signature = 'ddd:value {domain} {name}';
+    protected $name = 'ddd:value';
 
     /**
      * The console command description.
@@ -20,7 +16,20 @@ class MakeValueObject extends DomainGeneratorCommand
      */
     protected $description = 'Generate a value object';
 
-    protected $type = 'ValueObject';
+    protected $type = 'Value Object';
+
+    protected function getArguments()
+    {
+        return [
+            ...parent::getArguments(),
+
+            new InputArgument(
+                'name',
+                InputArgument::REQUIRED,
+                'The name of the value object',
+            ),
+        ];
+    }
 
     protected function getRelativeDomainNamespace(): string
     {