+++ tests/Generator/MakeBaseModelTest.php
@@ -1,25 +1,42 @@
 <?php
 
 use Illuminate\Support\Facades\Artisan;
+use Illuminate\Support\Facades\Config;
 
-it('can generate domain base model', function () {
+it('can generate domain base model', function ($domainPath, $domainRoot) {
+    Config::set('ddd.paths.domains', $domainPath);
+
     $modelName = 'BaseModel';
     $domain = 'Shared';
 
-    $expectedModelPath = base_path(implode('/', [
-        config('ddd.paths.domains'),
+    $expectedPath = base_path(implode('/', [
+        $domainPath,
         $domain,
         config('ddd.namespaces.models'),
         "{$modelName}.php",
     ]));
 
-    if (file_exists($expectedModelPath)) {
-        unlink($expectedModelPath);
+    if (file_exists($expectedPath)) {
+        unlink($expectedPath);
     }
 
-    expect(file_exists($expectedModelPath))->toBeFalse();
+    expect(file_exists($expectedPath))->toBeFalse();
 
     Artisan::call("ddd:base-model {$domain} {$modelName}");
 
-    expect(file_exists($expectedModelPath))->toBeTrue();
-});
+    expect(file_exists($expectedPath))->toBeTrue();
+
+    $expectedNamespace = implode('\\', [
+        $domainRoot,
+        $domain,
+        config('ddd.namespaces.models'),
+    ]);
+
+    expect(file_get_contents($expectedPath))->toContain("namespace {$expectedNamespace};");
+})->with('domainPaths');
+
+it('shows meaningful hints when prompting for missing input', function () {
+    $this->artisan('ddd:base-model')
+        ->expectsQuestion('What is the domain?', 'Shared')
+        ->assertExitCode(0);
+})->ifSupportsPromptForMissingInput();