+++ tests/TestCase.php
@@ -29,6 +29,9 @@ class TestCase extends Orchestra
             fn (string $modelName) => 'Lunarstorm\\LaravelDDD\\Database\\Factories\\'.class_basename($modelName).'Factory'
         );
 
+        File::deleteDirectory(resource_path('stubs/ddd'));
+        File::deleteDirectory(resource_path('config/ddd'));
+
         $this->beforeApplicationDestroyed(function () {
             File::cleanDirectory(app_path());
             File::deleteDirectory(base_path('Custom'));